+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [1.0.1] - Unreleased
-- 
+- Add master and domain CSV exports to the analytics generator and update generated files.
 
 ## [1.0.0]
 - Add new documentation files and configure pubspec for API reference links and dartdoc inclusion.